+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [Unreleased]
 
+### Added
+
+- Add support for reading ktlint version from `ktlint-plugins.properties` file [#1001](https://github.com/JLLeitschuh/ktlint-gradle/pull/1001)
+  - Enables automatic synchronization of ktlint version between IntelliJ plugin and Gradle builds
+  - When `ktlint-plugins.properties` exists in project root with `ktlint-version` property, that version is used as default
+  - Explicit version configuration in build script still takes precedence
+
 ## [14.0.1] - 2025-11-10
 
 - Update build to work with gradle 9.1 and Java 25 [#962](https://github.com/JLLeitschuh/ktlint-gradle/pull/962)